이동 로직 dtime 버그 1차 수정
attack에도 추가 예정

+++ YaMang/YaMangServer/YaMangServer/MovePosition.h
@@ -22,7 +22,7 @@ private:
 
 	
 	ULONGLONG		m_StartedTime = 0;
-	float			m_MovingTime = 0.0f;
+	ULONGLONG		m_MovingTime = 0;
 	PositionInfo	m_MovingRoute;
 };
